Add Vazir font support for the theme
Added a condition in base html file to check for variable THEME_FONT and
import the respective CSS to include the font. default font is
BMitra(Used when not specified in pelicanconf.py) however the said
variable can be defined from the pelicanconf.py file.

once set, the font changes to Vazir as expected.